feat!: Extract theme into an object passable to the config as `theme`
A theme gets the documentation-ready AST and uses a function passed to it in onBuild to "build" the output file structure.

With serveDynamic, this gets put in a "virtual" file system, with buildStatic, into the real file system.

BREAKING CHANGE: serveDynamic no longer renders the templates "on request", but instead once when called

Closes: #7